requirements.txt
# Core FastAPI Stack (compatible with Python 3.13)
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Pydantic v1 (no Rust requirement)
pydantic==1.10.13
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0

# Computer Vision (Task 3 dependencies)
ultralytics==8.0.196
opencv-python-headless==4.9.0.80
pillow==10.1.0
pandas==2.0.3
numpy==1.24.3
pyarrow==14.0.1
